+++ cli/CHANGELOG.md
@@ -7,6 +7,43 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0](https://github.com/tangle-network/gadget/compare/cargo-tangle-v0.1.3...cargo-tangle-v0.2.0) - 2024-10-24
+
+### Added
+
+- *(cargo-tangle)* [**breaking**] rename "gadget" to "blueprint" ([#384](https://github.com/tangle-network/gadget/pull/384))
+
+### Other
+
+- Event Flows for Tangle ([#363](https://github.com/tangle-network/gadget/pull/363))
+
+## [0.1.3](https://github.com/tangle-network/gadget/compare/cargo-tangle-v0.1.2...cargo-tangle-v0.1.3) - 2024-10-23
+
+### Added
+
+- *(cli)* support custom repo and path sources ([#360](https://github.com/tangle-network/gadget/pull/360))
+
+### Fixed
+
+- *(sdk)* updated keystore support and fixes ([#368](https://github.com/tangle-network/gadget/pull/368))
+- load substrate signer from env correctly ([#349](https://github.com/tangle-network/gadget/pull/349))
+
+### Other
+
+- release ([#378](https://github.com/tangle-network/gadget/pull/378))
+- release ([#362](https://github.com/tangle-network/gadget/pull/362))
+- update to latest changes in tangle ([#367](https://github.com/tangle-network/gadget/pull/367))
+- Event Workflows (phase 1: Custom listeners) ([#359](https://github.com/tangle-network/gadget/pull/359))
+- Event Listener Upgrade + Wrapper Types + sdk::main macro ([#333](https://github.com/tangle-network/gadget/pull/333))
+- update naming ([#343](https://github.com/tangle-network/gadget/pull/343))
+- release ([#314](https://github.com/tangle-network/gadget/pull/314))
+
+## [0.2.0](https://github.com/tangle-network/gadget/compare/cargo-tangle-v0.1.2...cargo-tangle-v0.2.0) - 2024-10-23
+
+### Fixed
+
+- *(sdk)* updated keystore support and fixes ([#368](https://github.com/tangle-network/gadget/pull/368))
+
 ## [0.1.2](https://github.com/tangle-network/gadget/releases/tag/cargo-tangle-v0.1.2) - 2024-09-30
 
 ### Other